% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_fn_utils.R
\name{db_auth}
\alias{db_auth}
\title{Authorize Dropbox}
\usage{
db_auth(db_token = file.path("inst", "auth", "db_token.rds"))
}
\arguments{
\item{db_token}{\code{(character)} path to Dropbox token saved as RDS}
}
\description{
Authorize Dropbox
}
