% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_QPR_tabitem.R
\name{mod_QPR_tabItem_ui}
\alias{mod_QPR_tabItem_ui}
\title{QPR_tabItem UI Function}
\usage{
mod_QPR_tabItem_ui(id, choices = NULL, date_choices = NULL)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{choices}{\code{(named list)} of arguments to \link[shinyWidgets]{pickerInput}.
 `FALSE` to omit the choice drop-down selector. - choices must be provided 
 for the picker to show options. Defaults will be used for omitted arguments 
 unless arguments are explicitly set to `NULL`. Defaults are as follows:
`list(
inputId = ns("region"),
label = "Select Region(s)",
choices = choices$choices, # 
options = shinyWidgets::pickerOptions(                                                  liveSearch = TRUE),
selected = NULL,
width = "70%"
)`
Default value for choices is found in *global.R* for each tabItem namespace.}

\item{date_choices}{\code{(logical)} whether to show a date range picker}
}
\description{
A shiny Module to generate the QPR tabitems.
}
\seealso{
Other QPR: 
\code{\link{mod_QPR_server}()}
}
\concept{QPR}
