% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{qpr_infobox}
\alias{qpr_infobox}
\title{qpr_infobox}
\usage{
qpr_infobox(
  .data,
  .replace = FALSE,
  title = "Average Score",
  color = "purple",
  value = .data$AvgScore,
  icon = shiny::icon("shopping-cart"),
  subtitle = paste(.data$Increased, "out of", .data$TotalHHs, "households served"),
  ...
)
}
\arguments{
\item{.data}{\code{(any)} Data to be passed in and used subsequent arguments}

\item{.replace}{\code{(logical)} whether to replace the default arguments with those supplied and eliminate the default arguments, or to replace existing defaults & and add additional args specified}

\item{title}{Title text.}

\item{color}{A color for the box. Valid colors are listed in
\link[shinydashboard]{validColors}.}

\item{value}{The value to display in the box. Usually a number or short text.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{subtitle}{Subtitle text (optional).}

\item{...}{
  Arguments passed on to \code{\link[shinydashboard:infoBox]{shinydashboard::infoBox}}
  \describe{
    \item{\code{width}}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}
    \item{\code{href}}{An optional URL to link to.}
    \item{\code{fill}}{If \code{FALSE} (the default), use a white background for the
content, and the \code{color} argument for the background of the icon. If
\code{TRUE}, use the \code{color} argument for the background of the
content; the icon will use the same color with a slightly darkened
background.}
  }}
}
\description{
Function to render infobox from default template for QPR tabitems
}
