% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_save.R
\name{sym2function}
\alias{sym2function}
\title{Convert the occurrence of a symbol in an R document to a function call by the same name}
\usage{
sym2function(x, file)
}
\arguments{
\item{x}{\code{(character)} All symbols to change}

\item{file}{\code{(character)} Path to file in which to change them}
}
\value{
\code{(character)} vector of text in the file with the new function calls
}
\description{
Useful to replace calls to an object by symbol name with an accessor function or reactive with the same name
}
