% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{ui_header_row}
\alias{ui_header_row}
\title{The UI Header output}
\usage{
ui_header_row(outputId = rlang::caller_env()$ns("header"))
}
\arguments{
\item{outputId}{\code{(namespaced character)} A character vector wrapped in `ns` from the parent environment.}
}
\value{
A fluidrow containing a minimizable box with the header
}
\description{
The UI Header output
}
