% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{ui_picker_project}
\alias{ui_picker_project}
\title{A pickerInput that provides active projects to select from}
\usage{
ui_picker_project(
  label = "Select Project",
  inputId = rlang::caller_env()$ns("project"),
  choices = projects,
  options = shinyWidgets::pickerOptions(liveSearch = TRUE, liveSearchStyle =
    "contains"),
  ...
)
}
\arguments{
\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{inputId}{\code{(namespaced character)} A character vector wrapped in `ns` from the parent environment.}

\item{choices}{List of values to select from. If elements of the
list are named then that name rather than the value is displayed to the user.}

\item{options}{List of options, see \link[shinyWidgets]{pickerOptions} for all available options.
To limit the number of selection possible, see example below.}

\item{...}{
  Arguments passed on to \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput}}
  \describe{
    \item{\code{selected}}{The initially selected value (or multiple values if \code{multiple = TRUE}).
If not specified then defaults to the first value for single-select lists
 and no values for multiple select lists.}
    \item{\code{multiple}}{Is selection of multiple items allowed?}
    \item{\code{choicesOpt}}{Options for choices in the dropdown menu.}
    \item{\code{width}}{The width of the input : 'auto', 'fit', '100px', '75\%'.}
    \item{\code{inline}}{Put the label and the picker on the same line.}
  }}
}
\value{
A select control that can be added to a UI definition.
}
\description{
A pickerInput that provides active projects to select from
}
